Allowlist stubby to be run within the path imposer
To support using google prod credentials for Google developers, we need
to be able to make an RPC call to exchange credentials. This RPC call
needs to be issued via stubby and hence allowlisting the stubby binary.

dwillemsen@ also pointed out that this should be less effective for
other tools running within the build since network access is disabled
for the other tools anyway via sandboxing.

Bug: b/222484146
Change-Id: I29aec9576e86b9ed3051df7943404abd478c549b